# [Ubuntu] Resolve TODOs, refactor /etc/environment parsing